Fix MSVC compiler warnings
reg.: conversion from 'size_t' to 'xxx', possible loss of data (C4267)
Plus some herder file cleanup

+++ importexport/musicxml/importmxmlpass2.cpp
@@ -4013,7 +4013,7 @@ void MusicXMLParserDirection::handleFraction()
 
       for (size_t n = 0; n < fracs.size(); n++) {
             if (rawWordsText.contains(fracs.at(n))) {
-                  int p = n <= 2 ? 0x00BC + n : 0x2150 + n - 3;
+                  size_t p = n <= 2 ? 0x00BC + n : 0x2150 + n - 3;
                   rawWordsText.replace(fracs.at(n), QString(char16_t(p)));
                   _wordsText = rawWordsText;
                   return;
